refactor: moved settings to utilties
