+++ src/sdk/services/blockchain/eosio/eosio.ts
@@ -1,7 +1,7 @@
 import { APIClient, FetchProvider } from '@greymass/eosio';
 import { GetInfoResponse } from '@greymass/eosio/src/api/v1/types';
 import fetch from 'cross-fetch';
-import { getSettings } from '../../../settings';
+import { getSettings } from '../../../util/settings';
 import { throwError, SdkErrors } from '../../../util/errors';
 
 let api: APIClient;
